requirements.txt
alembic==1.4.2
aniso8601==8.0.0
astroid @ file:///C:/ci/astroid_1592481955828/work
bcrypt==3.1.7
certifi==2020.6.20
cffi==1.14.1
click==7.1.2
colorama==0.4.3
dnspython==2.0.0
email-validator==1.1.1
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-JWT==0.3.2
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.10
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.19.1
Pillow==7.2.0
pycparser==2.20
PyJWT==1.4.2
pylint @ file:///C:/ci/pylint_1592487534522/work
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
scipy==1.5.2
six==1.15.0
SQLAlchemy==1.3.18
toml @ file:///tmp/build/80754af9/toml_1592853716807/work
typed-ast==1.4.1
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.11.2
WTForms==2.3.3
